Fix construction of BIP34 coinbase.
The expected coinbase for BIP34 blocks is constructed by serialising a
script with the block height as integer.  In the end, this leads to the
code in CScriptNum::serialize in the Bitcoin Core codebase.

Specifically, for positive numbers (as in the case of a block height),
the rule [1] is that another zero byte has to be pushed in case the last
byte had the highest bit set (since that bit is used to indicate the sign).

[1] https://github.com/bitcoin/bitcoin/blob/e117cfe45eee9169409e74a44ef4a866be25bc35/src/script/script.h#L351