add rainforest algo (#29)
A quick benchmark shows 1795 kH/s on a core i7/4790K at 4000 MHz and 795 kH/s on a Odroid-C2 ARM board at 1752 MHz lacking crypto extensions. Pretty nice for such a small device. (MikeMurdo)

It is very impressive on a Raspberry-Pi 3B+, it gives 668 kH/s, or roughly 1/3 of my skylake 4 GHz for 1/30 of the power usage and 1/18 of the price! Test report here : https://www.linkedin.com/pulse/more-efficient-mining-raspberry-pi-julien-delorme . The skylake also outperforms the GPU! (jdelorme3)

FWIW I just tested on NanoPI-Fire3 at 8*1.6 GHz and it's way faster (1825 kH/s). Apparently the algo makes use of crypto extensions which are enabled on this board but are not on raspi. It's fun to see a $35 device beat a $500 PC (wtarreau)

From https://github.com/bschn2/rainforest